Update rep DB for latest changes, e.g. Luther Strange

+++ database/populate_reps.py
@@ -4,13 +4,16 @@ import requests
 from database import db
 from database import db_models
 
+USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/56.0.2924.87 Safari/537.36'
+HEADERS = {'User-Agent': USER_AGENT}
+
 def main():
     populate_senators()
     populate_house_reps()
     db.session.commit()
 
 def populate_senators():
-    resp = requests.get('http://www.senate.gov/general/contact_information/senators_cfm.xml')
+    resp = requests.get('http://www.senate.gov/general/contact_information/senators_cfm.xml', headers=HEADERS)
     root = etree.fromstring(resp.content)
     for member in root.xpath('//member'):
         db_rep = db_models.Rep(
@@ -33,15 +36,16 @@ def populate_senators():
                 .filter(db_models.Rep.state_code == db_rep.state_code) \
                 .filter(db_models.Rep.chamber == db_models.Rep.Chamber.SENATE) \
                 .filter(db_models.Rep.status != db_models.Rep.Status.ACTIVE) \
-                .one()
-            db.session.delete(rep_to_delete)
-            db.session.add(db_rep)
-            print 'Deleting senator (%s) with status "%s": %s %s' % (rep_to_delete.state_code, rep_to_delete.status,
-                rep_to_delete.first_name, rep_to_delete.last_name)
+                .first()
+            if rep_to_delete:
+                db.session.delete(rep_to_delete)
+                print 'Deleting senator (%s) with status "%s": %s %s' % (rep_to_delete.state_code, rep_to_delete.status,
+                    rep_to_delete.first_name, rep_to_delete.last_name)
             print 'Adding new senator (%s): %s %s' % (db_rep.state_code, db_rep.first_name, db_rep.last_name)
+            db.session.add(db_rep)
 
 def populate_house_reps():
-    resp = requests.get('http://clerk.house.gov/xml/lists/MemberData.xml')
+    resp = requests.get('http://clerk.house.gov/xml/lists/MemberData.xml', headers=HEADERS)
     root = etree.fromstring(resp.content)
     for member in root.xpath('//members/member'):
         member_info = member.xpath('./member-info')[0]
